feat(undo/redo): fixed `groupBy` & `ignoreInitialState` issues
Refs: #43
